pyproject.toml
[project]
name = "auto-validator"
requires-python = "==3.11.*"
version = "0"
dependencies = [
    "Django~=4.2.4",
    "django-constance>=4.1.2",
    "django-cors-headers~=4.2.0",
    "django-environ~=0.11.2",
    "django-extensions==3.2.3",
    "django-probes==1.7.0",
    "django-debug-toolbar==4.1.0",
    "django-structlog[celery]==8.0.0",
    "celery~=5.3.1",
    "gunicorn==20.1.0",
    "psycopg[binary]~=3.1.19",
    "redis~=4.6.0",
    "sentry-sdk==1.3.0",
    "ipython~=8.14.0",
    "nox==2023.4.22",
    "django-fingerprint-rt~=0.1.0",
    "djangorestframework>=3.15.2",
    "django-storages[s3]>=1.14.4",
    "drf-spectacular>=0.27.2",
    "django-bootstrap-v5>=1.0.11",
    "discord.py==2.4.0",
    "python-dotenv>=1.0.1",
    "paramiko>=3.5.0",
    "scp>=0.15.0",
    "bittensor-cli>=8.1.1",
    "bittensor-wallet>=2.0.1",
    "bittensor>=8.1.1",
    "gitpython>=3.1.43",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
    'pytest>=8',
    'pytest-django>=4.8',
    'pytest-xdist>=3',
    'freezegun>=1.5',
]
lint = [
    "ruff>=0.6",
    "codespell[toml]",
]
type_check = [
    "django-stubs[compatible-mypy]",
    "djangorestframework-stubs[compatible-mypy]",
    "mypy>=1",
    "types-freezegun>=1",
    "types-python-dateutil>=2.8.2",
    "types-requests>=2.27",
]

[tool.ruff]
src = ["app/src"]
line-length = 120

[tool.ruff.lint]
# TODO add D
select = [
    "E", "F", "I", "UP", "S",
    "TCH005",
]
# TODO: remove E501 once docstrings are formatted
ignore = [
    "D100", "D105", "D107", "D200", "D202", "D203", "D205", "D212", "D400", "D401", "D415",
    "D101", "D102","D103", "D104", # TODO remove once we have docstring for all public methods
    "E501", # TODO: remove E501 once docstrings are formatted
    "S101", "S108", "S603", "S607",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/test*/**.py" = ["D", "F403", "F405", "S106", "S311"]

[tool.codespell]
skip = '*.min.js,*.lock,*/monitoring_certs/*'
ignore-words-list = 'datas'
